Rename usage_instructor to usage_util
