+++ modi/__main__.py
@@ -8,7 +8,7 @@ from getopt import getopt, GetoptError
 import modi
 
 from modi.util.debugging_util import Debugger
-from modi.util.usage_instructor import UsageInstructor
+from modi.util.usage_util import UsageInstructor
 from modi.util.tutorial_util import Tutor
 from modi.util.inspection_util import Inspector
